Move handlers for specific elements into sub-namespaces
