+++ src/Element/Document/InvalidateDocumentListener.php
@@ -1,6 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace Neusta\Pimcore\HttpCacheBundle\Element;
+namespace Neusta\Pimcore\HttpCacheBundle\Element\Document;
+
+use Neusta\Pimcore\HttpCacheBundle\Element\ElementInvalidationEvent;
+use Neusta\Pimcore\HttpCacheBundle\Element\ElementType;
 
 final class InvalidateDocumentListener
 {
