+++ src/Element/Object/TagObjectListener.php
@@ -1,32 +1,32 @@
 <?php declare(strict_types=1);
 
-namespace Neusta\Pimcore\HttpCacheBundle\Element;
+namespace Neusta\Pimcore\HttpCacheBundle\Element\Object;
 
 use Neusta\Pimcore\HttpCacheBundle\Cache\CacheTag;
 use Neusta\Pimcore\HttpCacheBundle\Cache\CacheTagCollector;
 use Neusta\Pimcore\HttpCacheBundle\CacheActivator;
-use Pimcore\Event\Model\AssetEvent;
+use Pimcore\Event\Model\DataObjectEvent;
 
-final class TagAssetListener
+final class TagObjectListener
 {
     public function __construct(
-        private readonly CacheActivator $cacheActivator,
         private readonly CacheTagCollector $cacheTagCollector,
+        private readonly CacheActivator $cacheActivator,
     ) {
     }
 
-    public function __invoke(AssetEvent $event): void
+    public function __invoke(DataObjectEvent $event): void
     {
         if (!$this->cacheActivator->isCachingActive()) {
             return;
         }
 
-        $asset = $event->getAsset();
+        $object = $event->getObject();
 
-        if ('folder' === $asset->getType()) {
+        if ('folder' === $object->getType()) {
             return;
         }
 
-        $this->cacheTagCollector->addTag(CacheTag::fromElement($asset));
+        $this->cacheTagCollector->addTag(CacheTag::fromElement($object));
     }
 }
